Changed ultrasonic output to float (untested) and added python capture script (unfinished)

+++ source/ultrasonic.h
@@ -8,6 +8,6 @@
 
 void initializeUltrasonic();
 
-long getUltrasonicOutput();
+float getUltrasonicOutput();
 
 #endif
